Add wall removability when DFS traversing
As the path is carved in the grid space, the walls are removed.
A significant bug was found where the rectangles were being
drawed at each visited grid space with a full stroke. This meant
that the walls 'seemed' to stay when they actually were properly
disappearing. At this stage, there is no recursive backtracking
so the maze stops forming once it hits a dead-end.