Add recursive backtracking to dfs algorithm
To add recursive backtracking, we simply need to push
the current cell to a stack during the normal traversal step
to new unvisited neighbors. Once we've hit a dead-end, we pop
the stack for whichever was the last cell pushed on, and then
recursively check for any unvisited neighbors we missed out on.
Interestingly, JS seems to implement its arrays with the same
functionality of a stack. Thus, we add only a couple lines of
code to implement the recursive backtracking (which sounds scary).